Moves the home assets to the home folder inside components.
