+++ index.html
@@ -5,7 +5,7 @@
 
   <meta charset="UTF-8">
 
-  <title>Angular 2 Hello World</title>
+  <title>Angular and Github API</title>
 
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css">
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bulma/0.3.1/css/bulma.min.css">
